refactor: ゲームユーティリティを分割
